requirements.txt
asgiref==3.6.0
black==22.12.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djhtml==1.5.2
filelock==3.8.2
identify==2.5.11
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.3
platformdirs==2.6.0
pre-commit==2.20.0
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
virtualenv==20.17.1
0x-contract-addresses==3.0.0
0x-contract-artifacts==3.0.0
0x-json-schemas==2.1.0
0x-order-utils==4.0.1
aiohttp==3.8.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==2.6.0
cbor2==5.4.3
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cytoolz==0.12.0
dacite==1.6.0
Deprecated==1.2.13
eth-abi
eth-account
eth-hash
eth-keyfile
eth-keys
eth-rlp
eth-typing
eth-utils
Flask
frozenlist==1.3.1
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.12.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.1
pycryptodome==3.15.0
pyee==9.0.4
pymerkle==2.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
requests==2.28.1
rlp==2.0.1
six==1.16.0
stringcase==1.2.0
thirdweb-contract-wrappers==2.0.4
thirdweb-eth-account==0.6.6
thirdweb-sdk==2.0.18
toolz==0.12.0
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.11
varint==1.0.2
web3==5.27.0
websockets==9.1
Werkzeug==2.2.0
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1